Updated frontend code.
